Simple error reporting for JSON decoding errors. Closes #23.
More user-friendly error reporting might be in order at some
point, or maybe not. The console based error message for this
occasion I decided to add, when I had to do `git bisect` just
to find out the reason why animation files didn't load after
I having added a new field to the Joint record, and which of
course didn't exist in files created prior to that change.

If I had the error on the console from day 0, I would have
seen the problem right away.